chore(1888): Addressed PR Comments
