+++ src/headerV2/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@buerokratt-ria/header",
-  "version": "0.1.49",
+  "version": "0.1.50",
   "description": "Generic Header component that would be injected as dependency.",
   "main": "index.ts",
   "scripts": {},
