+++ src/headerV2/src/components/DomainSelectorBar/index.tsx
@@ -7,23 +7,15 @@ import {getWidgetData, updateUserSelection} from '../../services/user';
 import {DomainSelection} from '../../types/widgetModels';
 import {ToastContextType} from '../../context/ToastContext';
 import {UserInfo} from '../../types/userInfo';
+import SelectedTick from './SelectedTick';
 import './DomainSelectorBar.scss';
 
-const SelectedTick: FC = () => (
-    <span className="domain-selector-bar__choice-check" aria-hidden>
-        <svg className="domain-selector-bar__tick-svg" viewBox="0 0 15 15" width="15" height="15" focusable="false">
-            <circle className="domain-selector-bar__tick-disc" cx="7.5" cy="7.5" r="7.5" />
-            <path className="domain-selector-bar__tick-mark" d="M4.3 7.8L6.7 10.1L10.7 5.2" />
-        </svg>
-    </span>
-);
-
-type SelectOption = {label: string; value: string; meta?: string};
+type SelectOption = {readonly label: string; readonly value: string; readonly meta?: string};
 
 type DomainSelectorBarProps = {
-    user: UserInfo;
-    toastContext: ToastContextType | null;
-    setUserDomains: (domains: string[]) => void;
+    readonly user: UserInfo;
+    readonly toastContext: ToastContextType | null;
+    readonly setUserDomains: (domains: string[]) => void;
 };
 
 function mapDomainSelections(domains: DomainSelection[]): {
